Handle absence of Git
When git has not be installed on the computer, we get a FileNotFoundError. Added an exception to handle this case, and also now log the get hash and navigate version number from the controller. Initially tried to do this from an __init__ file, which calls the _commit.py file, but I think this takes place before the logger is set up and therefore, of course, it does not log.