Update histogram.py
Now put the histogram in its own thread with a lock. There have been some complaints that the histogram causes latency, so hopefully this fixes it. If not, the next step will be to put a check box for enabling/disabling the histogram, and we would store the state of the checkbox in the experiment.yaml file.